requirements.txt
appdirs==1.4.1
begins==0.9
packaging==16.8
pyparsing==2.1.10
six==1.10.0
